Added CIDR block overlap check
